% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{betaHMM}
\alias{betaHMM}
\alias{betaHMM,matrix-method}
\alias{betaHMM-methods}
\alias{betaHMM,data.frame-method}
\title{HMM for beta valued DNA data}
\usage{
betaHMM(object, ...)

\S4method{betaHMM}{matrix}(object, M = 3, N = 4, R = 2, seed = NULL, ...)

\S4method{betaHMM}{data.frame}(object, M = 3, N = 4, R = 2, seed = NULL, ...)
}
\arguments{
\item{object}{Annotated data to be clustered. Maybe provided as a matrix or dataframe.}

\item{...}{Extra arguments}

\item{M}{Number of methylation states to be identified in a DNA sample.}

\item{N}{Number of patients or DNA sample replicates collected for each
treatment group.}

\item{R}{Number of treatment groups (For. eg: Benign and Tumour).}

\item{seed}{Seed to allow for reproducibility (default = NULL).}
}
\value{
An S4 object of class \code{betaHMMResults}.
}
\description{
This is the primary user interface for the \code{betaHMM} package.
Generic S4 methods are implemented to perform a homogeneous hidden Markov model for the beta valued DNA
             methylation data.
The supported classes are \code{matrix} and \code{data.frame}.
The output of \code{betaHMM} is an S4 object of class \code{betaHMMResults}.
}
\author{
Koyel Majumdar
}
\keyword{methods}
