% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{annotatedData}
\alias{annotatedData}
\alias{K}
\alias{N}
\alias{R}
\alias{A}
\alias{phi}
\alias{llk}
\alias{tau}
\alias{hidden_states}
\alias{annotatedData,betaHMMResults-method}
\alias{K,RangedSummarizedExperiment-method}
\alias{K,betaHMMResults-method}
\alias{K,NULL-method}
\alias{N,RangedSummarizedExperiment-method}
\alias{N,betaHMMResults-method}
\alias{N,NULL-method}
\alias{R,RangedSummarizedExperiment-method}
\alias{R,betaHMMResults-method}
\alias{R,NULL-method}
\alias{A,RangedSummarizedExperiment-method}
\alias{A,betaHMMResults-method}
\alias{A,NULL-method}
\alias{tau,RangedSummarizedExperiment-method}
\alias{tau,betaHMMResults-method}
\alias{tau,NULL-method}
\alias{phi,RangedSummarizedExperiment-method}
\alias{phi,betaHMMResults-method}
\alias{phi,NULL-method}
\alias{llk,RangedSummarizedExperiment-method}
\alias{llk,betaHMMResults-method}
\alias{llk,NULL-method}
\alias{hidden_states,RangedSummarizedExperiment-method}
\alias{hidden_states,betaHMMResults-method}
\alias{hidden_states,NULL-method}
\title{Accessors for the assigned cluster labels of a coseqResults object.}
\usage{
annotatedData(object)

K(object)

N(object)

R(object)

A(object)

phi(object)

llk(object)

tau(object)

hidden_states(object)

\S4method{annotatedData}{betaHMMResults}(object)

\S4method{K}{RangedSummarizedExperiment}(object)

\S4method{K}{betaHMMResults}(object)

\S4method{K}{`NULL`}(object)

\S4method{N}{RangedSummarizedExperiment}(object)

\S4method{N}{betaHMMResults}(object)

\S4method{N}{`NULL`}(object)

\S4method{R}{RangedSummarizedExperiment}(object)

\S4method{R}{betaHMMResults}(object)

\S4method{R}{`NULL`}(object)

\S4method{A}{RangedSummarizedExperiment}(object)

\S4method{A}{betaHMMResults}(object)

\S4method{A}{`NULL`}(object)

\S4method{tau}{RangedSummarizedExperiment}(object)

\S4method{tau}{betaHMMResults}(object)

\S4method{tau}{`NULL`}(object)

\S4method{llk}{RangedSummarizedExperiment}(object)

\S4method{llk}{betaHMMResults}(object)

\S4method{llk}{`NULL`}(object)

\S4method{phi}{RangedSummarizedExperiment}(object)

\S4method{phi}{betaHMMResults}(object)

\S4method{phi}{`NULL`}(object)

\S4method{hidden_states}{RangedSummarizedExperiment}(object)

\S4method{hidden_states}{betaHMMResults}(object)

\S4method{hidden_states}{`NULL`}(object)
}
\arguments{
\item{object}{a \code{betaHMMResults} object.}
}
\description{
The counts slot holds the count data as a matrix of non-negative integer
count values, one row for each observational unit (gene or the like), and one
column for each sample.
}
