% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{betaHMMResults-class}
\alias{betaHMMResults-class}
\alias{betaHMMResults}
\title{betaHMMResults object and constructor}
\usage{
betaHMMResults(SummarizedExperiment, annotatedData)
}
\arguments{
\item{SummarizedExperiment}{a \code{RangedSummarizedExperiment} of \code{betaHMM} results}

\item{annotatedData}{The annotated data passed as an input argument to the betaHMM package.}
}
\value{
a betaHMMResults object
}
\description{
\code{betaHMMResults} is a subclass of \code{RangedSummarizedExperiment},
used to store the betaHMM results as well as the annotated data useful
 for plotting.
}
\details{
This constructor function would not typically be used by "end users".
This simple class extends the \code{RangedSummarizedExperiment} class of the
SummarizedExperiment package
to allow other packages to write methods for results
objects from the betaHMM package. It is used by
to wrap up the results table.
}
