% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{dmcResults-class}
\alias{dmcResults-class}
\alias{dmcResults}
\title{dmcResults object and constructor}
\usage{
dmcResults(SummarizedExperiment)
}
\arguments{
\item{SummarizedExperiment}{a \code{RangedSummarizedExperiment} of
\code{dmcResults} results.}
}
\value{
a \code{\link[betaHMM:dmcResults]{dmcResults}} object
}
\description{
\code{dmcResults} is a subclass of \code{RangedSummarizedExperiment},
used to store the DMCs identified.
}
\details{
This constructor function would not typically be used by "end users".
This simple class extends the \code{RangedSummarizedExperiment} class of the
SummarizedExperiment package
to allow other packages to write methods for results
objects from the
\code{\link[betaHMM:dmc_identification]{dmc_identification}} function.
It is used by
to wrap up the results table.
}
