% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{dmr_identification}
\alias{dmr_identification}
\alias{dmr_identification,betaHMMResults-method}
\alias{dmr_identification-methods}
\title{DMR identification from DMCs}
\usage{
dmr_identification(object, ...)

\S4method{dmr_identification}{betaHMMResults}(object, diff_meth_cluster = 1, ...)
}
\arguments{
\item{object}{a \code{betaHMMResults} object.}

\item{...}{extra arguments}

\item{diff_meth_cluster}{The clusters identified as differentially methylated}
}
\value{
A dataframe containing the following columns:
\itemize{
\item start_CpG - The starting CpG site in the particular DMR
\item end_CpG -  The ending CpG site in the particular DMR
\item DMR_size - Number of CPG sites identified in the DMR
\item chr_dmr - The chromosome corresponding to the CpG sites in the DMR.
\item map_start - MAPINFO of starting CpG site in the particular DMR
\item map_end -  MAPINFO of the ending CpG site in the particular DMR}
}
\description{
Function to identify the DMRs for the DMCs identified using BHMM
}
\details{
Function to identify the DMRs for the DMCs identified using BHMM
}
\seealso{
\code{\link{betaHMM}}
}
