% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-functions.R
\name{dmr_identification_run}
\alias{dmr_identification_run}
\title{DMR identification from DMCs}
\usage{
dmr_identification_run(object, diff_meth_cluster, ...)
}
\arguments{
\item{object}{A \code{\link[betaHMM:betaHMM]{betaHMM}} object.}

\item{diff_meth_cluster}{The clusters identified as differentially methylated}

\item{...}{extra arguments}
}
\value{
A dataframe containing the following columns:
\itemize{
\item start_CpG - The starting CpG site in the particular DMR
\item end_CpG -  The ending CpG site in the particular DMR
\item DMR_size - Number of CPG sites identified in the DMR
\item chr_dmr - The chromosome corresponding to the CpG sites in the DMR.
\item map_start - MAPINFO of starting CpG site in the particular DMR
\item map_end -  MAPINFO of the ending CpG site in the particular DMR}
}
\description{
Function to identify the DMRs for the DMCs identified using BHMM
}
\details{
Function to identify the DMRs for the DMCs identified using BHMM
}
