% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{dmrResults-class}
\alias{dmrResults-class}
\alias{dmrResults}
\title{dmrResults object and constructor}
\usage{
dmrResults(SummarizedExperiment)
}
\arguments{
\item{SummarizedExperiment}{a \code{dmrResults}
results.}
}
\value{
a  \code{\link[betaHMM:dmrResults]{dmrResults}} object
}
\description{
\code{dmrResults} is a subclass of \code{RangedSummarizedExperiment},
used to store the DMRs identified.
}
\details{
This constructor function would not typically be used by "end users".
This simple class extends the \code{RangedSummarizedExperiment} class of the
SummarizedExperiment package
to allow other packages to write methods for results
objects from the
\code{\link[betaHMM:dmr_identification]{dmr_identification}} function.
It is used by
to wrap up the results table.
}
