% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot-functions.R
\name{plot}
\alias{plot}
\alias{plot,betaHMMResults-method}
\alias{plot-methods}
\title{Visualize results from betaHMM clustering}
\usage{
plot(x, ...)

\S4method{plot}{betaHMMResults}(
  x,
  what = c("fitted density", "kernel density", "uncertainty"),
  treatment_group = NULL,
  title = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{"betaHMMResults"}}

\item{...}{Other graphics parameters.}

\item{what}{The different plots that can be obtained are either
"fitted density","kernel density" or
"uncertainty" (default = "fitted density").}

\item{treatment_group}{The names of the different treatment groups.
If no value is passed then default values of sample names, e.g. Sample 1,
Sample 2, etc are used as legend text (default = NULL).}

\item{title}{The title that the user wants to display.
If no title is to be displayed the default is "NULL".}
}
\value{
This function displays the following plots as requested by the user:
\itemize{
\item fitted density estimates - Plot showing the fitted density estimates of the clustering solution under the optimal model selected.
\item kernel density estimates - Plot showing the kernel density estimates of the clustering solution under the optimal model selected.
\item uncertainty -  A boxplot showing the uncertainties in the optimal clustering solution.
}
}
\description{
Plot a betaHMMResults object.
}
\author{
Koyel Majumdar
}
