% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_thresholds.R
\name{plot_thresholds}
\alias{plot_thresholds}
\title{The plot for threshold indentification}
\usage{
plot_thresholds(x, data, title = NULL)
}
\arguments{
\item{x}{A \code{\link[betaHMM:threshold_identification]{threshold_identification}} object.}

\item{data}{A dataframe of dimension \eqn{C \times N} containing methylation
values for \eqn{C} CpG sites from one treatment group where
DNA samples are either collected from \eqn{N} patients
or has \eqn{N} replicates for the treatment group and this
dataset was passed as an argument to the
\code{\link[betaHMM:threshold_identification]{threshold_identification}} function.}

\item{title}{The title for the plot.}
}
\value{
The plot for the estimated shape parameters and threshold for the methylation states.
}
\description{
The plot for threshold indentification
}
\details{
The plot for threshold indentification
}
