% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward.R
\name{backward}
\alias{backward}
\title{Backward algorithm}
\usage{
backward(probabilities, trained_params)
}
\arguments{
\item{probabilities}{Probability calculated based on the initialised
model parameters.}

\item{trained_params}{A list containing the initialised parameters for the
 BHMM (initial distribution, transition matrix,
shape parameters for the observation data distribution).}
}
\value{
A list containing:
\itemize{
   \item log_beta - The backward probabilities.
   \item scaled_logL - The log-likelihood calculated using the backward
                       probabilities.
   }
}
\description{
Compute the backward probabilities for the BHMM model.
}
\details{
Computes the backward probabilities for the number of hidden states
         using the initial distribution, transition matrix and distribution
         parameters.
}
