% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaumWelch.R
\name{BaumWelch}
\alias{BaumWelch}
\title{Estimation Using the Baum-Welch Algorithm}
\usage{
BaumWelch(data, trained_params, M, N, R, n.iter = 100, seed = NULL)
}
\arguments{
\item{data}{A dataframe of dimension \eqn{C \times (N \times R)} containing
methylation values for \eqn{C} CpG sites from \eqn{R}
treatment groups each having \eqn{N} replicates or each collected
from \eqn{N} patients.}

\item{trained_params}{A list containing the initialised model parameters:
\itemize{
   \item A - The transition matrix for the BHMM model.
   \item tau - The initial distribution for the BHMM model.
   \item phi - The shape parameters for the observation sequence data
               in the BHMM model.
   }}

\item{M}{Number of methylation states to be identified in a DNA sample.}

\item{N}{Number of patients or DNA sample replicates collected for each
treatment group.}

\item{R}{Number of treatment groups (For. eg: Benign and Tumour).}

\item{n.iter}{Number of iterations for algorithm convergence.}

\item{seed}{Seed to allow for reproducibility (default = NULL).}
}
\value{
A list containing:
\itemize{
   \item A - The transition matrix for the BHMM model.
   \item tau - The initial distribution for the BHMM model.
   \item phi - The shape parameters for the observation sequence data
               in the BHMM model.
   \item log_vec - A vector containing the log-likelihood values calculated
                   for each iteration of the algorithm.
   \item z - A matrix of dimension \eqn{C \times K} containing the posterior
             probability of each CpG site belonging to each of the
             \eqn{K} clusters.
   }
}
\description{
The Baum-Welch algorithm is used to estimate the parameters of
             the beta distributed hidden Markov model (BHMM).
}
\details{
The Baum-Welch algorithm is used to estimate the parameters of
             the beta distributed hidden Markov model (BHMM).
}
