% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{forward}
\alias{forward}
\title{Forward algorithm}
\usage{
forward(probabilities, trained_params)
}
\arguments{
\item{probabilities}{Probability calculated based on the initialised
model parameters.}

\item{trained_params}{A list containing the initialised parameters for the
 BHMM (initial distribution, transition matrix,
shape parameters for the observation data distribution).}
}
\value{
A list containing:
\itemize{
   \item log_alpha - The forward probabilities.
   \item scaled_logL - The log-likelihood calculated using the forward
                       probabilities.
   }
}
\description{
Compute the forward probabilities for the BHMM model.
}
\details{
Computes the forward probabilities for the number of hidden states
         using the initial distribution, transition matrix and distribution
         parameters.
}
