% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot-functions.R,
%   R/plotDMR-functions.R, R/plot_thresholds.R
\name{plot}
\alias{plot}
\alias{plot,betaHMMResults-method}
\alias{plot-methods}
\alias{plot,dmcResults-method}
\alias{plot,threshold_Results-method}
\title{Visualize results from betaHMM clustering}
\usage{
plot(x, ...)

\S4method{plot}{betaHMMResults}(
  x,
  chromosome = NULL,
  what = c("fitted density", "kernel density", "uncertainty"),
  treatment_group = NULL,
  AUC = NULL,
  uncertainty_threshold = 0.2,
  title = NULL,
  ...
)

\S4method{plot}{dmcResults}(
  x,
  start_CpG = NULL,
  end_CpG = NULL,
  treatment_group = NULL,
  N = NULL,
  title = NULL,
  ...
)

\S4method{plot}{threshold_Results}(x, title = NULL, ...)
}
\arguments{
\item{x}{An object of
class \code{threshold_Results}}

\item{...}{Extra plot parameters.}

\item{chromosome}{The chromosome number for which the plot is to be displayed}

\item{what}{The different plots that can be obtained are either
'fitted density','kernel density' or
'uncertainty' (default = 'fitted density').}

\item{treatment_group}{The names of the different treatment groups to be
displayed in the plot.
If no value is passed then the treatment group names from the
\code{betaHMM} function are used.}

\item{AUC}{The AUC values for that chromosome.}

\item{uncertainty_threshold}{The uncertainty threshold value used for DMC
identification.}

\item{title}{The title for the plot.}

\item{start_CpG}{The IlmnID of starting CpG site.}

\item{end_CpG}{The IlmnID of ending CpG site/ the total number
of CpGs to be plotted excluing the starting CpG site.}

\item{N}{Number of replicates in each treatment group.}
}
\value{
This function displays the following plots as requested by the user:
\itemize{
\item fitted density estimates - Plot showing the fitted density estimates
of the clustering solution under the optimal model selected.
\item kernel density estimates - Plot showing the kernel density estimates
of the clustering solution under the optimal model selected.
\item uncertainty -  A boxplot showing the uncertainties in the
hidden state estimation.
}

This function displays the DMCs and DMRs plot from
the \code{\link[betaHMM:dmcResults]{dmcResults}} object.

The plot for the estimated shape parameters and threshold
for the methylation states in a single DNA treatment condition.
}
\description{
Plot a \code{betaHMMResults} object.

Plot a \code{\link[betaHMM:dmcResults]{dmcResults}} object.

The plot for threshold identification in a single
DNA treatment condition.
}
\details{
Visualize the 3 methylation states in DNA replicates collected from
one condition

Plot a \code{threshold_Results} object.
}
\examples{
## Use simulated data for the betaHMM workflow example
set.seed(12345)
library(betaHMM)

## read files
data(sample_methylation_file)
head(sample_methylation_file)
data(sample_annotation_file)
head(sample_annotation_file)
## Run betaHMM function
beta_out <- betaHMM(sample_methylation_file, sample_annotation_file,
                    M = 3, N = 4, R = 2,
                    parallel_process = FALSE, seed = 12345,
                    treatment_group = c("Benign","Tumour"))

## Run dmc_identification function
dmc_out <- dmc_identification(beta_out)
dmc_df <- assay(dmc_out)

# Run dmr_identification function
dmr_out <- dmr_identification(dmc_out, parallel_process = FALSE)
#
# Plot functions
# Get the AUC values calculated for each hidden state
AUC_chr <- AUC(dmc_out)

## plot the fitted density estimates
plot(beta_out, chromosome = "1", what = "fitted density", AUC = AUC_chr)

## Plot the DMR graph
plot(dmc_out, start_CpG = "cg14817997", end_CpG = 8)
## Use simulated data for the betaHMM workflow example
set.seed(12345)
library(betaHMM)

## read files
data(sample_methylation_file)
head(sample_methylation_file)
data(sample_annotation_file)
head(sample_annotation_file)
## Run betaHMM function
beta_out <- betaHMM(sample_methylation_file, sample_annotation_file,
                    M = 3, N = 4, R = 2,
                    parallel_process = FALSE, seed = 12345,
                    treatment_group = c("Benign","Tumour"))

## Run dmc_identification function
dmc_out <- dmc_identification(beta_out)
dmc_df <- assay(dmc_out)

# Run dmr_identification function
dmr_out <- dmr_identification(dmc_out, parallel_process = FALSE)
#
# Plot functions
# Get the AUC values calculated for each hidden state
AUC_chr <- AUC(dmc_out)

## plot the fitted density estimates
plot(beta_out, chromosome = "1", what = "fitted density", AUC = AUC_chr)

## Plot the DMR graph
plot(dmc_out, start_CpG = "cg14817997", end_CpG = 8)
}
\author{
Koyel Majumdar
}
