% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{threshold_Results-class}
\alias{threshold_Results-class}
\alias{threshold_Results}
\title{threshold_Results object and constructor}
\usage{
threshold_Results(SummarizedExperiment, annotatedData)
}
\arguments{
\item{SummarizedExperiment}{a \code{RangedSummarizedExperiment} of
\code{threshold_Results} object.}

\item{annotatedData}{The annotated data passed as an input argument to
the \code{threshold_identification} function.}
}
\value{
a \code{\link[betaHMM:threshold_Results]{threshold_Results}} object
}
\description{
\code{threshold_Results} is a subclass of \code{RangedSummarizedExperiment},
used to store the threshold_identification results as well as the annotated
 data useful for plotting.
}
\details{
This constructor function would not typically be used by "end users".
This simple class extends the \code{RangedSummarizedExperiment} class of the
SummarizedExperiment package
to allow other packages to write methods for results
objects from the
\code{\link[betaHMM:threshold_identification]{threshold_identification}}
function. It is used by
to wrap up the results table.
}
